Change id to BIGINT

+++ src/main/resources/sql/shopr_ddl.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user (
-  id                INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+  id                BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   name              VARCHAR(255) NOT NULL,
   first_name        VARCHAR(255) NOT NULL
 );
